refactor(step3):   코드정리
Scroe 커스텀 익셉션 사용
player 패키지 분리
