+++ src/main/java/bowling/controller/BowlingGameRun.java
@@ -1,7 +1,7 @@
 package bowling.controller;
 
 import bowling.domain.BowlingGame;
-import bowling.domain.Player;
+import bowling.domain.player.Player;
 import bowling.domain.frame.Frame;
 import bowling.domain.frame.Frames;
 import bowling.domain.point.Point;
