+++ src/main/java/bowling/domain/player/Player.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.player;
 
 import java.util.Optional;
