+++ src/main/java/bowling/domain/score/ScroeAccessDenyException.java
@@ -1,0 +1,8 @@
+package bowling.domain.score;
+
+public class ScroeAccessDenyException extends IllegalArgumentException {
+
+    public ScroeAccessDenyException(String msg){
+        super(msg);
+    }
+}
